relocate renameChannel function
